# IntIterator and LongIterator implement OfInt,OfLong