--- a/RoaringBitmap/src/main/java/org/roaringbitmap/BatchIntIterator.java
+++ b/RoaringBitmap/src/main/java/org/roaringbitmap/BatchIntIterator.java
@@ -38,7 +38,7 @@ public class BatchIntIterator implements IntIterator {
   }
 
   @Override
-  public int next() {
+  public int nextInt() {
     return buffer[i++];
   }