--- a/RoaringBitmap/src/main/java/org/roaringbitmap/ImmutableBitmapDataProvider.java
+++ b/RoaringBitmap/src/main/java/org/roaringbitmap/ImmutableBitmapDataProvider.java
@@ -313,7 +313,7 @@ public interface ImmutableBitmapDataProvider {
 
     @Override
     public int nextInt() {
-      return iterator.next();
+      return iterator.nextInt();
     }
 
     @Override