--- a/RoaringBitmap/src/main/java/org/roaringbitmap/IntIterator.java
+++ b/RoaringBitmap/src/main/java/org/roaringbitmap/IntIterator.java
@@ -4,13 +4,15 @@
 
 package org.roaringbitmap;
 
+import java.util.PrimitiveIterator.OfInt;
+
 /**
  * A simple iterator over integer values.
  * Using an IntIterator instead of Java's Iterator&lt;Integer&gt;
  * avoids the overhead of the Interger class: on some tests,
  * IntIterator is nearly twice as fast as Iterator&lt;Integer&gt;.
  */
-public interface IntIterator extends Cloneable {
+public interface IntIterator extends Cloneable, OfInt {
   /**
    * Creates a copy of the iterator.
    * 
@@ -26,6 +28,6 @@ public interface IntIterator extends Cloneable {
   /**
    * @return next integer value
    */
-  int next();
+  int nextInt();
 
 }