--- a/RoaringBitmap/src/main/java/org/roaringbitmap/IntIteratorFlyweight.java
+++ b/RoaringBitmap/src/main/java/org/roaringbitmap/IntIteratorFlyweight.java
@@ -64,7 +64,7 @@ public class IntIteratorFlyweight implements PeekableIntIterator {
   }
 
   @Override
-  public int next() {
+  public int nextInt() {
     int x = iter.nextAsInt() | hs;
     if (!iter.hasNext()) {
       ++pos;