--- a/RoaringBitmap/src/main/java/org/roaringbitmap/ReverseIntIteratorFlyweight.java
+++ b/RoaringBitmap/src/main/java/org/roaringbitmap/ReverseIntIteratorFlyweight.java
@@ -68,7 +68,7 @@ public class ReverseIntIteratorFlyweight implements IntIterator {
 
 
   @Override
-  public int next() {
+  public int nextInt() {
     final int x = iter.nextAsInt() | hs;
     if (!iter.hasNext()) {
       --pos;