--- a/RoaringBitmap/src/main/java/org/roaringbitmap/buffer/BufferBitSetUtil.java
+++ b/RoaringBitmap/src/main/java/org/roaringbitmap/buffer/BufferBitSetUtil.java
@@ -115,7 +115,7 @@ public class BufferBitSetUtil {
     }
     final IntIterator it = bitmap.getIntIterator();
     while (it.hasNext()) {
-      int val = it.next();
+      int val = it.nextInt();
       if (!bitset.get(val)) {
         return false;
       }