--- a/RoaringBitmap/src/main/java/org/roaringbitmap/buffer/BufferIntIteratorFlyweight.java
+++ b/RoaringBitmap/src/main/java/org/roaringbitmap/buffer/BufferIntIteratorFlyweight.java
@@ -69,7 +69,7 @@ public class BufferIntIteratorFlyweight implements PeekableIntIterator {
   }
 
   @Override
-  public int next() {
+  public int nextInt() {
     int x = iter.nextAsInt() | hs;
     if (!iter.hasNext()) {
       ++pos;