--- a/RoaringBitmap/src/main/java/org/roaringbitmap/buffer/BufferReverseIntIteratorFlyweight.java
+++ b/RoaringBitmap/src/main/java/org/roaringbitmap/buffer/BufferReverseIntIteratorFlyweight.java
@@ -73,7 +73,7 @@ public class BufferReverseIntIteratorFlyweight implements IntIterator {
 
 
   @Override
-  public int next() {
+  public int nextInt() {
     final int x = iter.nextAsInt() | hs;
     if (!iter.hasNext()) {
       --pos;