--- a/RoaringBitmap/src/main/java/org/roaringbitmap/longlong/LongIterator.java
+++ b/RoaringBitmap/src/main/java/org/roaringbitmap/longlong/LongIterator.java
@@ -3,27 +3,18 @@
  */
 package org.roaringbitmap.longlong;
 
+import java.util.PrimitiveIterator.OfLong;
+
 /**
  * A simple iterator over long values. Using an IntIterator instead of Java's Iterator&lt;Long&gt;
  * avoids the overhead of the Long class: on some tests, LongIterator is nearly twice as fast as
  * Iterator&lt;Long&gt;.
  */
-public interface LongIterator extends Cloneable {
+public interface LongIterator extends Cloneable, OfLong {
   /**
    * Creates a copy of the iterator.
    * 
    * @return a clone of the current iterator
    */
   LongIterator clone();
-
-  /**
-   * @return whether there is another value
-   */
-  boolean hasNext();
-
-  /**
-   * @return next long value
-   */
-  long next();
-
 }