--- a/RoaringBitmap/src/test/java/org/roaringbitmap/TestIntIteratorFlyweight.java
+++ b/RoaringBitmap/src/test/java/org/roaringbitmap/TestIntIteratorFlyweight.java
@@ -29,7 +29,7 @@ public class TestIntIteratorFlyweight {
       if (!(size < values.length)) {
         values = Arrays.copyOf(values, values.length * 2);
       }
-      values[size++] = ints.next();
+      values[size++] = ints.nextInt();
     }
     return Ints.asList(Arrays.copyOf(values, size));
   }