--- a/jmh/src/jmh/java/org/roaringbitmap/realdata/wrapper/EwahIteratorWrapper.java
+++ b/jmh/src/jmh/java/org/roaringbitmap/realdata/wrapper/EwahIteratorWrapper.java
@@ -16,7 +16,7 @@ final class EwahIteratorWrapper implements BitmapIterator {
   }
 
   @Override
-  public int next() {
+  public int nextInt() {
     return iterator.next();
   }