--- a/jmh/src/jmh/java/org/roaringbitmap/realdata/wrapper/RoaringBitmapWrapper.java
+++ b/jmh/src/jmh/java/org/roaringbitmap/realdata/wrapper/RoaringBitmapWrapper.java
@@ -23,7 +23,7 @@ final class RoaringBitmapWrapper implements Bitmap {
 
   @Override
   public int last() {
-    return bitmap.getReverseIntIterator().next();
+    return bitmap.getReverseIntIterator().nextInt();
   }
 
   @Override