--- a/jmh/src/jmh/java/org/roaringbitmap/realdata/wrapper/RoaringIteratorWrapper.java
+++ b/jmh/src/jmh/java/org/roaringbitmap/realdata/wrapper/RoaringIteratorWrapper.java
@@ -17,7 +17,7 @@ final class RoaringIteratorWrapper implements BitmapIterator {
 
   @Override
   public int next() {
-    return iterator.next();
+    return iterator.nextInt();
   }
 
 }