--- a/jmh/src/jmh/java/org/roaringbitmap/realdata/wrapper/RoaringIteratorWrapper.java
+++ b/jmh/src/jmh/java/org/roaringbitmap/realdata/wrapper/RoaringIteratorWrapper.java
@@ -16,7 +16,7 @@ final class RoaringIteratorWrapper implements BitmapIterator {
   }
 
   @Override
-  public int next() {
+  public int nextInt() {
     return iterator.nextInt();
   }