This makes it easier to use in standard library situations that expect and OfInt or OfLong. The downside is that while practicaly compatible due to autoboxing existing users of the next() method might be slowed down if they do not migrate to nextLong().

### SUMMARY
Since java 8 there has been an interface OfInt and OfLong that is similar to the IntIterator and LongIterator in ideas.
However instead of next returning the primitive, next returns an autoboxed one, with new methods nextInt/Long to return a primitive.

This pull request changes uses the two primitive Iterators to OfInt and OfLong. This makes it easier to interoperate with other code that expects these two interfaces. But with a significant downside that existing of `next()` may have a significant regression in performance.

I personally don't think this should be merged before the 0.10.x release series but wanted to open this pull request for discussion if and when such a patch should be merged.

Procedure to make this patch is trivial. Use rename re-factoring of the existing next to nextInt/Long methods then add the OfInt/Long to the existing Int/LongInterface. 

### Automated Checks

- [x] I have run `./gradlew test` and made sure that my PR does not break any unit test.
- [x] I have run `./gradlew checkstyleMain` or the equivalent and corrected the formatting warnings reported.