@blacelle 2022-11-01
Relates with https://github.com/RoaringBitmap/RoaringBitmap/pull/461

I may require the same API change, hence package change.