@lemire 2022-11-01
> But with a significant downside that existing of next() may have a significant regression in performance.

Is that true?

Certainly, that may happen if 'scalarisation' fails, but I am interested in examples where the iterator would be such that scalarisation would not happen.

Here is what I got (quick and dirty) on my laptop...


Before:
```
$ ./jmh/run.sh 'BitmapIteratorBenchmark'                                                 1

# JMH version: 1.23
# VM version: JDK 11.0.17, OpenJDK 64-Bit Server VM, 11.0.17+8
# VM invoker: /Library/Java/JavaVirtualMachines/temurin-11.jdk/Contents/Home/bin/java
# VM options: -XX:-TieredCompilation
# Warmup: 5 iterations, 10 s each
# Measurement: 5 iterations, 10 s each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Average time, time/op
# Benchmark: org.roaringbitmap.iteration.BitmapIteratorBenchmark.backwards
# Parameters: (density = 0.1)

# Run progress: 0,00% complete, ETA 00:33:20
# Fork: 1 of 1
# Warmup Iteration   1: 13,219 us/op
# Warmup Iteration   2: 13,207 us/op
# Warmup Iteration   3: 12,133 us/op
# Warmup Iteration   4: 12,281 us/op
# Warmup Iteration   5: 12,438 us/op
Iteration   1: 12,102 us/op
Iteration   2: 12,524 us/op
Iteration   3: 12,823 us/op
Iteration   4: 12,310 us/op
Iteration   5: 12,323 us/op


Result "org.roaringbitmap.iteration.BitmapIteratorBenchmark.backwards":
  12,416 ±(99.9%) 1,046 us/op [Average]
  (min, avg, max) = (12,102, 12,416, 12,823), stdev = 0,272
  CI (99.9%): [11,370, 13,463] (assumes normal distribution)
```




After:
```
$ ./jmh/run.sh 'BitmapIteratorBenchmark'                                                                                    1
# Run progress: 0,00% complete, ETA 00:33:20
# Fork: 1 of 1
# Warmup Iteration   1: 13,272 us/op
# Warmup Iteration   2: 13,235 us/op
# Warmup Iteration   3: 12,224 us/op
# Warmup Iteration   4: 12,211 us/op
# Warmup Iteration   5: 12,296 us/op
Iteration   1: 12,413 us/op
Iteration   2: 12,271 us/op
Iteration   3: 12,371 us/op
Iteration   4: 12,335 us/op
Iteration   5: 12,301 us/op


Result "org.roaringbitmap.iteration.BitmapIteratorBenchmark.backwards":
  12,338 ±(99.9%) 0,217 us/op [Average]
  (min, avg, max) = (12,271, 12,338, 12,413), stdev = 0,056
  CI (99.9%): [12,122, 12,555] (assumes normal distribution)

```



I have not run the full benchmark, we test forward and backward, and with different densities. However, as you can see, there is no regression in at least one case.


Could someone volunteer to benchmark this?