@lemire 2022-11-01
My naive expectation is that if you take an integer, you box it, return the box, and then immediately unbox it to an integer... then the JIT will just optimize away the boxing. At least this ought to be true for 'some' versions of Java.

This presentation makes me hopeful : 

https://www.youtube.com/watch?v=p1MhRBYS-k0

In any case, it is an empirical issue. We can test it out.