@JervenBolleman 2022-11-01
@lemire I assumed as much in most cases. But the comment on IntIterator suggests that there are cases where scalarization used to fail and `Iterator<Integer>` was slower than `IntIterator` by a factor 2x. 

>  Long class: on some tests, LongIterator is nearly twice as fast as Iterator<Long>. 

With basically identical text in the IntIterator.
I have not checked which test these are and if that is still the case on any jvm that is supported.