@JervenBolleman 2022-11-03
Well I ran into an issue :( when adapting the existing `RealDataBenchmarkIterate.iterate` jmh usecase.

> java.lang.AbstractMethodError: Receiver class io.druid.extendedset.intset.ConciseSet$BitIterator does not define or inherit an implementation of the resolved method 'int nextInt()' of interface io.druid.extendedset.intset.IntSet$IntIterator.

this might lead to more issues downstream than I expected.