@lemire 2022-11-03
Why would you need to adapt it?