@JervenBolleman 2022-11-03
Oh to see the difference of running the test with `nextInt()` and the same with `next()`. A before after could have worked but then I would need to be careful to see if the new test uses `next()` all the way through. A side by side test seemed more straightforward to quickly do over my lunch break :)