@JervenBolleman 2022-11-03
Ah yes the druid problem is from [here](https://github.com/apache/druid/blob/c5fcc03bdfd50f2624ba0a3d39f0d5e5e856dee/extendedset/src/main/java/org/apache/druid/extendedset/intset/IntSet.java#L63). Which could impact users which would use the new `nextInt()` method on their implementations unless special care is taken as in the [wrapper](https://github.com/RoaringBitmap/RoaringBitmap/pull/591/commits/5a348b693ec83106f8861a714624beb87309b341#diff-b643f761e7c5a207013d67605191777ef759a136a1e20842ffc5d9d52b45a075). 

This is different from me adapting the exiting JMH benches other than those surfacing this issue.