@lemire 2022-11-03
Yes. That's an example where this PR would break existing code. :-(